@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.1+20
 
 environment:
-  sdk: '>=3.2.0 <4.0.0'
+  sdk: '>=3.5.0 <4.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -37,27 +37,27 @@ dependencies:
   card_settings: ^3.5.0
   shared_preferences: ^2.2.2
   flutter_launcher_icons: ^0.14.1
-  select_searchable_list: ^0.1.05
-  dropdown_search: ^5.0.6
-  fl_chart: ^0.66.1
+  select_searchable_list: ^0.2.2
+  dropdown_search: ^6.0.1
+  fl_chart: ^0.69.0
   receive_sharing_intent: ^1.6.7
   googleapis: ^13.0.0
   extension_google_sign_in_as_googleapis_auth: ^2.0.12
   google_sign_in: ^6.2.1
   download_assets: ^3.1.1
-  tar: ^1.0.3
+  tar: ^2.0.0
   envied: ^0.5.4+1
   envied_generator: ^0.5.3
   build_runner: ^2.4.8
   logger: ^2.1.0
-  worker_manager: ^6.3.2
-  connectivity_plus: ^2.3.8
+  worker_manager: ^7.2.2
   flutter_settings_ui: ^3.0.0
-  window_manager: ^0.3.8
+  window_manager: ^0.4.2
   http: any
   googleapis_auth: any
   system_info2: ^4.0.0
   url_launcher: ^6.3.0
+  connectivity_plus: ^6.0.5
 
 dev_dependencies:
   flutter_test:
@@ -68,8 +68,8 @@ dev_dependencies:
   # activated in the `analysis_options.yaml` file located at the root of your
   # package. See that file for information about deactivating specific lint
   # rules and activating additional ones.
-  flutter_lints: ^3.0.1
-  ffigen: ^11.0.0
+  flutter_lints: ^5.0.0
+  ffigen: ^14.0.1
   msix: ^3.16.8
 
 # For information on the generic Dart part of this file, see the
