@@ -45,11 +45,15 @@ inline double Exponential(double x) {
 }
 
 inline double PolyLog(double x) {
-  return pow(-log(x), 1.5);
+  // Avoid log(0), in case the power below made the number 0 with -Ofast.
+  return pow(-log(std::max(x, 1E-30)), 1.5);
 }
 
 inline double ExponentialTimesPolyLog(double x) {
   assert(x >= 0 && x <= 1);
+  if (x <= 1E-18) {
+    return -std::numeric_limits<double>::max();
+  }
   double power = 1.7;
   return -pow(Exponential(pow(x, power)) * PolyLog(pow(x, power)), 1 / power);
 }
@@ -57,7 +61,7 @@ inline double ExponentialTimesPolyLog(double x) {
 class ProbCombiner {
  public:
   constexpr ProbCombiner(double (*function)(double x)): function_(function) {
-    assert(f(0) == -std::numeric_limits<double>::infinity());
+    assert(f(0) <= -std::numeric_limits<double>::max());
     assert(f(1) == 0);
   }
 
@@ -66,7 +70,7 @@ class ProbCombiner {
   constexpr double inverse(double y) {
     assert(y <= 0);
     // Used to avoid infinity when computing the inverse.
-    if (function_(1E-14) > y) {
+    if (function_(1E-14) >= y) {
       return 0;
     }
     return Inverse(function_, y, 1E-14, 1);
